Make it explicit that Rhombus discussion is open to everyone
Someone on Slack asked if feedback from industrial programmers would be useful for Rhombus designers. The answer is "of course!" but I realized that the Rhombus README doesn't really say anywhere who the intended participants in Rhombus's design are. This pull request fixes that and makes it clear that Rhombus's design is intended to be a community effort that takes place in the open, not behind closed doors.