reoragnize operators and transformers, use binding spaces for extension
Try out binding spaces as a way to allow existing bindings to be
enriched.

It works less well to use binding spaces as the exlcusive way to
distinguish, say, expression versus binding-position transformers.
Expression bindings really want to live in the default space, and then
a binding-space-specific lookup will fall back to the default space
and get the wrong kind of transformer. But it works well to cotninue
to have different transforer representations (via structure
properties) and use binding spaces as an option or fallback mechanism
for extending an existing binding.

Also, change the operator representation one more time.