add keyword to shrubbery notation and change keyword syntax
Use 'identifier' to mean a keyword. A single quote was otherwise
unused and allowed in an operator name, but a quote mark is strange as
an operator (for a non-Lisp syntax), and it's helpful to have a
notation of keyword that is separate from identifiers.

The syntax for keyword arguments, etc., now can be `'identifier':`
instead of `identifier:>`. Using `:` has the further advantage of
creating a block, which is more consistent and more flexible.