change binding-macro protocol again
Use a success--fail continuation style, defunctionalized at the
compile-time level so that matches can be represented as syntax, and
parameterized over an `if` form so that the expansion can be either
nested binding to implement a `match` search (without resorting to
`call/cc`) or an inline sequence of definitions for `value` (using
`error` for match failure).

Obvious in retrospect.