further generalize `val` right-hand side handling
Instead of an ad hoc rule that recognizes a handful of forms, force
enforestations of a block on the right-hand side that contains a
single group where the group does not parse as a definition. The
single-group, non-definition constraint is still somewhat ad hoc, but
less than the old rule.