add `......` for tail repetition, make `...` in `?` bind to lists
When a pattern variable in `?` is followed by `...`, bind the variable
to a list of syntax, instead of a syntax object that is a
parenthesized sequence. Tail repetition via `......` binds to a syntax
object with parentheses as before.

Also, adjust the shrubbery parser to introduce syntax pairs sometimes
that can house scope adjustments and avoid related quadratic behavior.