add `@` notation
The `@` notation here is almost the same as `#lang at-exp`, but
adjusted for shrubbery notation. A term

  @<command>[<arg>, ...]{<body>}

is equivalent to

  <command>(<arg>, ..., [<parsed-body-elem>, ...])

If delimiting is needed around <command>, use `«` and `»` immediately
after `@` instead of `|` and `|`; one or more terms in the <command>
group are allowed, and they are effectively spliced into the context.
General <body> brackets that start `|`....`{` and end `}`....`|` still
use `|`. The comment form is `@//` instead of `@;` and only works
within a <body>.