change interaction parsing
Consider single-line input complete as long as it's balanced and have
a `:` outside of an opener-closer pair. That way, for example,
entering `1+2` produces `3` immediately. For multi-line input, require
a blank line.

The `#%top-interaction` form splices the content of any top-level
block that's its own group into the top-level sequence. So, you can
always start input with a `:` to put it in multi-line mode.