change range indentation
When a range is selected, try to indent it deterministically. If a
deterministic indentation exists and is different from the current
state, then use it. (A determistic choice should happen any time the
range is armored.) If a deterministc indentation exists but is
not a change, cycle through shifts of the whole block.