make a `$` escape along in a group context match the whole group
Based on a suggestion from @AlexKnauth on Discord, an `$id` escape in
a syntax pattern stands for either a term or a group, depending on the
context.

For example, instead of

```
def '(thunk: $body ...; ...):
  '(fun () $body ...; ...)
```

you could write just

```
def '(thunk: $body; ...):
  '(fun () $body; ...)
```

Both of the above macro definitions work. In the first variant, `$body
...` does not have `$body` by itself, so the `body` stands for a term.

To match a single term alone within its group, annotate the pattern
variable with the `Term` syntax class, as opposed to the `Group`
syntax class:

```
'($(x $: Term)): '(10)
```

This initial choise of `$:` for a syntax-class annotation is based on
rhombus-brainstorming/discussions#194.

A group syntax object that's matched to multiple terms can only be
used in a template in a group context. So, if `body` is bound to a
group syntax object with multiple terms, `'(x $body)` triggers an
error.

Single-term group syntax objects are interchangeable with term syntax
objects. That rule mostly makes things "just work", such as when you
match input that you know for some reason has a single item in its
group and then use that input in a larger group. That situation
happens with the low-level binding macro example in the Rhombus
documentation, for example.

My initial experience with this is good. It's possible that
gropp-versus-term inference plus autoconversion is a little too
magical, though. Also, `$(id $: Term)` is noisy compared to just `$id`
for matching a term alone in its group.