shrubbery: adjust parsing of trailing `.` for a number
Treat `.` as the last character of a number only when it cannot be
treated instead as the start of a subsequent multi-character operator.
This change allows sequences like `..` and `.=` to be used as
operators immediately after a number, but without disallowing
something like `1.` in other contexts as `1.0`.

For example, `1..2` is the same as `1 .. 2`, and `1.=2` is the same as
`1 .= 2`, but `1.+2` is `1. + 2` (because `.+` is not an operator).

The rule is not perfect, because someone could write `1.*x` intending
`1. * x`, but `1.*x` is `1 .* x`. (The case of `+` in `1.+2` is
different, because there's a rule about `+`, `-`, `.`, and `/` at the
end of an operator to avoid ambiguity related to number and comment
starters.)