Updates to the equality RRFI
Narrow the scope of the document to the two-level scheme, with the
content pertaining to the primitive predicate extracted and soon to be
added to a separate document.

Simplify the organization, put the terminology at the end for
reference instead of having it at the beginning.

Incorporates feedback from @rocketnia and others.