remove `.....` while adding tail rule for `...`
The `......` pattern and template operator formerly served the same
role as using `.` in a Racket syntax pattern: matching/representing
the tail of a sequence without touching the sequence elements. Now
that a Rhombus syntax object can reresent a sequence of elements as a
group or a multi-group sequence, we can use that representation for a
tail match, and there are less representation issues with different
kinds of sequences. So, change the meaning of `...` in a tail position
to produce and consume those representations.

In some cases, currently, a tail `$x ...` or `$x; ...` needs to be
written instead as `$(x :: Term) ...` or `$(x :: Group); ...` to
disable tail matching and form a plain list of syntax objects instead
of a tail syntax object. Even that problem can be avoided (in the long
run) by using stream operators instead of list operators. Meanwhile,
`...` in a template works with coercion rules to do the right thing
without extra hassle.