fix phase for pattern and template parsing
Previously, in something like

  expr.macro '$a <> $b':
    ....

the pattern was being parsed as a phase-0 term but it should instead
be parsed as a phase-1 term. This will especially matter when new
syntax classes can be defined.

Meanwhile, in

  expr.rule '$a <> $b':
    ....

the pattern is still parsed as a phase-0 term, even though the parse
expansion is used as a phase-1 term. That's almost certainly the wrong
idea; the intent of `expr.rule` is to allow a macro to use `$` and
`...` without importing into phase 1, but this mismatch is going to
cause trouble with syntax classes.