for: drop `~and`, allow folder with `~into`, add `~do`
Instead of `~and` as a variant of `~each` that must follow an `~each`,
change `~each` to allow an immediate block containing binding
sequences to run in parallel.

Allow `~into` at the end of a `for` body to supply a reducer (former a
"folder"). A reducer can be place at the beginning or the end,
whichever way seemed clearer for a given use.

Add `~do` to enable macros that generate `for` bodies, analogous to
`#:splice`. The names `each`, etc., are bound as for_clause forms that
just expand to `~each`, etc.; alternatively, view `~each` as a
shorthand for `~do each`. The rationale here is that recognizing just
`each` by itself means that there would be a position where
expressions and for_clause forms overlap, which is asking for trouble,
and requiring `~do` keeps the two kinds of positions separate.
Meanwhile, the predefined `for` controls are common enough that it's
worth having shorthands.

This commit also fixes Scribble problems in `rhombus` and
`rhombusblock` where binding-space inference had been disabled.