improve implementation of `.`ed names
Instead of having a `List` annotation macro recognize `.` and then
`of`, for example, create a name root. To support a use of just the
root name like `List`, use a new enforest protocol that can redirect
the root to another compile-time binding.

Besides making the implementation of dotted names more regular, this
approach can work with Scribble to make a name like `List.of`
hyperlinked independent of `List`.