fill in some list, map, set, and array methods
Add `List.array`, `Map.count`, `Set.count`, and `Array.length`, each
of which can be applied to a value using `.length`, etc., as in
`[1, 2, 3].length`. Also add `Array.make`.

This is not so much meant to fill out the datatype story as prototype
and demonstrate how methods could be done. There's a lot of
boilerplate in the current implementation; it will certainly get
abstracted in some Rhombus-level interface, and maybe also internally,
although it's not initially clear how much it matters for a small
number of built-in datatypes.

Also, change `Map.make` and `Set.make` to `MutableMap` and `MutableSet`
(as suggested by @AlexKnauth).