improve some error reporting
As a compromise for better error reporting, the Rhombus expander now
disallows a reference to identifiers before they ae bound. Currently,
due to the way partial expansion works, this might only rule out
sequences that would otherwise provoke a use-before-definition error
at run time. If macros force some expansion early under a delaying
form, this might rule out programs that could run --- but probably the
forced expansion creates other trouble. This doesn't work for
top-level evaluation, so it's enabled only for module expansion, and
we may have to revisit.

A lot of the clean up in error reporting is related to defining
macros. The changes here provide a way to bind `$` as a macro and
to create a rule that expands to a literal `$` or `...`.