add `&&` and `||` as binding operators
For now, `||` doesn't bind anything from either its right or left
bindings. (This is in contrast to `match`, where `or` requires the
same bindings from each side; requiring the same bindings does not
seem practical here, but providing a way to propagate values seems
workable.)

Also, unbreak `class` in a non-module context, and clean up matching
on `$`, `...`, and `::` (use corerct space and phase) in syntax
patterns and templates.