try out an `import` and `export` shorthand
Allowing a `:` to be left out doesn't save much typing at all, but I
find myself omitting the `:` all the time (and then getting an error).
Maybe it's better to have that error and keep things slightly simpler,
or maybe it's better to allow a shorthand. After having tried the
first for a while, trying the second option for a while.