allow `&` and `...` in the middle of calls and constructions
Instead of requiring `&` or `...` in a call or construction to be at
the end, allow them anywhere in the list/map/set/argument sequence,
and allow multiple of them. Bindings sill require a single instance at
the end.

Allow `~&` anywhere in a call, but restrict it to a single instance.

Preserve the order of evaluation in all cases, including the order of
a `~&` case relative to others.

Ensure that something like `[&l]` always generates a list, etc.