revise custom constructor/binding/annotation and internal
Allow a subclass of a class with a custom constructor to use a default
constructor. The default assumes that the superclass constructor
accepts arguments consistent with its `field_spec`s.

Add support for `private` in a `field_spec`, so that it's possible to
add new immutable fields plus a custom constructor without foorcing
changes in subclasses.

Get rid of the wacky syntactic currying and non-hygienic `super`
binding for custom bindings and annotations. Instead, customizations
can use the name bound by `internal` plus a `&&` binding or annotation
operator.

Fix the `internal` name as a constructor to be consistent with a
custom constructor in the superclass. In particular, it has the same
calling convention as `super` within a constructor --- but `super` is
still needed, because that potentially instantiates a subclass, while
the internal name always instanties the class its associated with.

Fix `internal` names to consistently provide access to private fields
and methods.