switch `for` form away from keywords
Based on experience with `class`, change `for` to use binding instead
of keywords for structural forms like `each` and `break_when`. Part of
the trade-off here is that names need to be more distinct to avoid
clause--expression ambiguity, however, so change `when` and `unless`
to `keep_when` and `skip_when`.