more fully distinguish properties from methods
Change the syntax of `property` declarations to reflect the different
way that they;re accessed and called, and bring the syntax of direct
use within a class or interface in line with the `.` syntax.

These changes also correct the handling of static information for
properties, so that an annotation can communicate static information
to a use of a property without interference from the assignment mode
of properties.