check up the `check` form and move it into `rhombus`
The `check` form corresponds to a check in RackUnit, but with a single
form that has a few modes, instead of many different `check` forms.
Future additions can support more of the test-case and test-suite
structure that RackUnit provides.

This commit also reorganizes to support implementing more of `rhombus`
in (pre-)Rhombus via `#lang rhombus/private/core`.