Add a basic `Equatable` interface
This interface doesn't provide much more than the bare bones of an equality API. All it provides is this:

```
interface Equatable:
  method equals(other) :: Boolean
  method hashCode() :: Fixnum
```

That's it. No support for a `recur` argument, distinguishing between `equal?` and `equal-always?`, or an equality key is provided. This is all that's needed to get Rhombus collections working. Objects that implement this interface will get an implementation of `prop:equal+hash`, but that implementation will not use the `recur` or `mode` arguments.