add special treatment of '$[]` escape in a template
A `$[]` escape is analogous to a `#,@` escape in S-expression
templates --- although it does build in a list construction, so
there's a closer equivalence between `#,@....` and `$[&....]`. The
syntax `$[]` is available because a list is not valid to replace a
non-splicing `$` escape.

Since template and expression repetitions are unified, `$[]` can also
achieve the same result as `~@` in S-expression templates.

 > def [x, ...] = [1, 2, 3]
 > '0 $['+', x] ...'
 '0 + 1 + 2 + 3'

That's not quite the same as `~@`, which would allow the `+` and `x`
combination to be written like `'+ $x'`. Using `&` and `unwrap_group`
is noisier but matches the expressiveness of `~@`:

 > '0 $[& '+ $x'.unwrap_group()] ...'
 '0 + 1 + 2 + 3'