change `$[]` to `$$`
On further reflection, relying on the default binding of `[]` for the
meaning of a `$[]` shrubbery seems like a bad idea. Avoiding an extra
operator binding was nicer, but it seems worth using up the operator
`$$` to avoid the pun and to have something that can directly
implement both `#,@` and `~@`. The new `$$` allows the escaped
expression to return either a list or a syntax object, where the
syntax object can be a group whose elements are spliced.

 > def [x, ...] = [1, 2, 3]
 > '0 $$[x, ...]'
 '0 1 2 3'
 > '0 $$['+', x] ...'
 '0 + 1 + 2 + 3'
 > '0 $$('+ $x') ...'
 '0 + 1 + 2 + 3'

The `rhombusblock` form in `scribble/rhombus` uses the newly available
`#,` instead of `$$`, which is a change that I was considering anyway.