add an intermediate step to the low-level binder protocol
The binder protocol splits definitions needed for matching from final
definitions that may be modified by `let`, and so the final
definitions must not refer to each other. Putting both the "commit"
work and the possibility of `let` adjustments in the same secod part,
however, turns out not to be general enough.

Specifically, and as an example, a repetition binding needs an
intermediate point the representation's backing (currently a list)
needs to be created, and then a repetition binding refers to that
backing.

The solution here is to add an intermediate step that is not often
needed, but available for binding forms that need an intermediate
binding that is computed only after committing to a match.

Closes #269