collapse `rule` and `macro` forms
Rename the plain `rule` form to `macro`, and otherwise collapse:

 * a `macro` form doesn't have to match a tail anymore;

 * if a macro wants to match the tail, it should use a pattern that
   ends with `$identifier ...` or the `$(~end)` pseudo-escape; and

 * a macro can return two syntax objects only when it has a
   tail-matching pattern.

The overall result is to make things much simpler: there is only
`macro`, most of the time a macro author can ignore the tail (because
most of the time it should just be preserved), and there's a smoother
transition between those two. It's even more rare that a macro author
would need to use `$(~end)`; the main uses would be to insist that no
further tail follows or to have a more complex repetition (not just an
identifier) to cover the tail.