change underlying mapping of Rhombus spaces
Instead of implementing different kinds of binding by using the
default space unless one of the `only` binding forms is used, always
bind in a specific space. This approach avoids the need for `only`
binding forms, and so we don't need a convention about when `only`
should be used.

Internally, the default space is still used, but as the specific space
for expressions. That choice means that an expression binding shadows
across all spaces, which was part of the original idea to avoid some
especially confusing binding interactions. For example, this approach
avoids a weird result when a repetition is meant to be shadowed by a
non-repetition:

    def [x, ...] = [1, 2, 3]
    begin:
      def x = 1
      [x, ...]

Using the default space for the inner `x` makes it shadow the
repetition binding and report an error about using `x` in a too-deep
repetition. If `x` were bound in a more specific underlying space,
then the result would see past the inner `x` binding and produce
`[1, 2, 3]`.

To further avoid confusion, a repetition binding creates an
accompanying binding in the expression space (to a form that always
errors); the extra binding makes the other nesting order between
expressions and repetitions also produce helpful errors. This choice
for repeptition-binding forms breaks the usual pattern, but it
reflects how repetition looks just like variable binding --- and
reflects an assumption that you never want to bind a name with
unrelated meanings in the expression and repetition spaces.

There's still some room for confusion with namespaces (whose uses also
overlap expression positions). For example, with

  class Posn(x, y)
  def p = Posn(x, y)

  begin:
    namespace p:
      def x = 10
      export: x

    p

it's likely surprising that the result is `Posn(1, 2)` and not an
error about misusing `p`. Meanwhile, `p.x` would access the exported
`x` from the namespace, not from the `Posn` instance, because
namespace paths take precedence. In other words, the inner `namespace
binding does not shadow the outer variable. It would be even weirder
if

  namespace p:
    export x
    def x = 0

  begin:
    class Posn(x, y)
    def p = Posn(1, 2)
    p.x

produced 0 instead of 1, but that's an example of the kind of
ambiguity that is prevented by using the default space for
expressions. In other words, expression bindings (which are by far the
most common) defend against weird interactions with other spaces, but
other spaces have to pick an interaction when they overlap with
expressions. Usually they should not do anything special, but there
are cases like repetitions where it's better to behave a little more
like before.

Overall, the old strategy of defaulting binding to the default space
avoided more ambiguities, but it also required more machinery and more
explanation. After trying out both approaches, this new one seems
better.

A lot of this commit cleans up sloppy handling of binding in patterns
that would need to be cleaned up either way. Using the default binding
space hid problems with the sloppy implementation, so in that way, the
old approach was worse. Related to that, literals in syntax patterns
now match symbolically, not by binding; to match by binding, you have
to use a `binding` form and specify the relevant space.

Another substantial part of the commit is adjusting the way that
namespaces work. Formerly, dotted path elements could be found in
either the default space or the context's space. Now, namespace
elements up to the last one are always in a namespace-specific space.
That choice surfaces the ability to define a name both as a namespace
and as other forms --- like the way that a class name is used as a
namespace as well as an expression, binding form, annotation form, and
more.