change macro pattern strategy again: no `~parsed` or `~end`, but `$()`
In the macro patterns '$a x', `$a x $b', and 'x $b', one or both
escapes are treated as standing for parsed form. If you want to match
a singe non-parsed thing as the only after the defined name, you need
to use a syntax class, as in `x $(b :: Term)`. Also, the pattern
implicitly allows additional termas after the patch, unless you use
`$()` at the end of the pattern.