consistent and liberal convention on `:` for various forms
When a form can have any number of subforms (e.g., `import`,
`implements`), the general approach is to expect a block after the
keyword. A `:` for a block is redundant and noisy for other forms that
take only a single subform (e.g., `extends` in a class, `internal`,
and an `~op_stx` option in an operator body).

To avoid having to think about which convention should apply (and
remember whether a form can have multiple subforms), make the `:`
optional in most cases. For forms that support multiple subforms,
omitting `:` means that the rest of the group is treated the same as a
single group in a block. For a form that allows only a single subform,
using `:` is sometimes constrained to a single group (or even single
identifier, such as for `internal`) in the block, and sometimes it
creates a body block (such as a block after `break_when`).

This isn't a general rule built into shrubbery notation. It's
effectively a convention about how many different patterns a macro
implementation should have, but exactly what makes sense will vary for
different forms.