`match` in `fun`, `opeperator`, and `macro` to group alternatives
In addition to

 fun | <case>
     | ...

allow

 fun <name> <maybe-res-annotation>:
   match | <case>
         | ...

so that the result annotation can be written just once. For `operator`
and macro-definition forms, precedence and associatvity to be shared
across cases can be written before `match`.

This is a slight abose of the `match` name and binding, since it's not
the `match` expression form (but meant to be suggestive of it).