change `expr_meta.Group` to `expr_meta.Parsed`
When using a syntax class like `expr_meta.Group`, it's easy to forget
to use the `parsed` field to get the parsed value, while the terms
directly associated with a match are rarely useful. Allow the "root"
value of a pattern variable to be swapped with a field, and use that
swapping to make `expr_meta.Parsed`. When a pattern variable is
annotated with the `expr_meta.Parsed` synatx class, its value is the
parsed result, while a `group` field holds the original terms.