rhombus/gui: use `obervable-view`
Where view contrutors take other views, support observables and
convert them to views using `observable-view` at the `gui-easy` level.

Remove the custom view conditionals, such as `View.if`. Also remove
`ListView`, because the same effect is available through an observable
whose value is a panel.

Simplify Checkbox and TabsPanel creation with a default action that
sets an obsevable, and report the observable through a property.