document naming convention and try to follow it
A new "Naming Conventions" section at the end of the manual describes
the naming conventions that are (intended to be) used in the current
`rhombus` APIs.

Notable changes:

 * Change `Maybe` to `maybe`, because it's an annotation constructor and
   does not refer to an interface (even in principle).

 * Change `StringViewAsString` to `StringView.to_string`, and similar for
   `rhombus/draw` converting annotations.

The name `StringView.to_string` is derived from Discord discussion
(adjusted here for newly documented conventions), but there was more
discussion there:

 * `StringView.something` makes sense if you think of it from the
   provider perspective, where it is satisfied by the same values as
   `StringView` but with some extra provisos, similar to the way that
   `List.of(...)` is satisfied by a list.

 * Starting with `String` instead of `StringView` could also make
   sense, especially from the receiver perspective, since the
   resulting value satisfies `String`.

To go along with the `StringView.to_string` annotation, there is a
`StringView.to_string` function, and a `to_string` works (statically
or dynamically) for anything satisfying `StringView`. Note that the
annotation `StringView.to_string` has parentheses, similar to the way
`List` as an annotation has no parentheses, while calling the function
`StringView.to_string` obviously needs parentheses, just like the
`List` constructor is called with parentheses. It's not clear that
this overloading of `to_string` is a good idea; one alternative would
be `as_string`, but then a programmers has to remember different names
for the same idea in different contexts.

In the case of `PointView.to_point` and similar, `to_point` works as a
method only for static dispatch on something with static information
from the `PointView` annotation. Having a `PointView.to_point`
function seems useful and worthwhile, but it's not clear that the
static-only method is a good idea.

The trick of adjusting error messages by renaming an annotation on
import is messy for the `PointView.to_point` function, which is
defined in the same module as `PointView`. A new `described_as`
binding operator makes it easy to provide an "annotation" shown in
error reporting (really: any shrubbery term sequence) that is separate
from the implementation of an annotation or binding.