add `assign.macro` for defining assignment operators
Support for assignment is built into `.`, `[]`, field access, and
proprty access. Previously, those places recognized `:=` specifically,
but now they can access operators defined with `assign.macro`.

An assignment macro can expand to a use of another assignment macro,
or it can use `assign_meta.unpack_left` to get an accessor function
and mutator function (and a name to use as the inferred name for the
right-hand side). It can then use the accessor and mutator in whatever
combination, such as using both for implement a C-like `+=` operator,
and return it by packing with `assign_meta.pack_assignment`.