add `Sequence`, `Sequenceable`, and `sequence`
The `Sequence` annotation is satisfied by any sequence, which
coincides with Racket's `sequence?`. The `Sequence.make` function is a
keywordized and more direct (removes a thunk layer) variant of
Racket's `make-do-sequence`, while the combination of
`Sequence.instantiable` and `Sequence.instantiate` supports the same
instantiation indirection as `make-do-sequence`.

Implementing the `Sequenceable` interface requires supplying a
`to_sequence` method that returns a sequence, possibly created by
`Sequence.make`.

The `sequence` class-clause form specializes `each` expansion in `for`
for a right-hand side that has the class-as-annotation's static
information (and ditto for interfaces). The `sequence` form is like
`define-sequence-syntax`, except that it essentially forces a `:do-in`
(or `#false`) result and doesn't expose a new identifier that might be
used as an expression. The specialization's result is translated to
`:do-in`.