add `namespace ~open`
Creating a namespace and immediately openeing it is useful for having
a mixture of definitions where only some are available outside the set
of definitions. An `~open` shortcut is especially useful for a
macro-generated namespace, where you want the opened names to have the
same visibility in terms of scopes that they would have with the
definitions spliced directly into a context.