fix merging `export` cases for the same identifier
For example, one `export` might be for an identifier at one space,
while another is for a different identifier in a different spaces, but
with the same symbolic name.

To support these combinations, the representation of a namespace is a
little more complicated, and `export` merging can be fairly
complicated. On the plus side, the new namespace representation
simplifies some import expansions, because extra rename transformers
are not needed.