improve range-shifting reindent for call arguments
Improve reindent for something like

```
 the_f(x,
  y,
  z,
  w)
```

where the lines for `y` through `w` are selected. The strategy of
checking whether the later lines can be tabbed under the first
line as-is doesn't work when the `(` says to tab further, but we
cna tweak the indentation search to stop before it goes past
the first selected line.