change `Name` to recognize dotted names; reorganize syntax meta
The main goal of this change is to help macros match and manipulate
dotted sequences. A pattern escape `$(x :: Name)` matches an
identifier, operator, or sequence of components separated by `.` ---
where all components are identifiers, except that the last one can be
a parenthesized operator. The `IdentifierName` syntax class is
similar, but it recognizes only identifiers, which is along the lines
of forms like `def` and `namespace` that bind only identifiers and not
operators.

These syntax classes have the kind `~sequence`, which means they can
be used in the middle of a group. They match a spliced sequence of
terms independent of binding. They also match eagerly (i.e., consuming
as many `.`s and elements as possible within the group), which is just
a consequence of how syntax matching is eager in general.

After matching a dotted path, the next step might be to look up
something about its binding. Support for that step leads to a larger
set of changes in this commit. Even merely comparing bindings may need
to consult compile-time values to resolve dotted names, which doesn't
work in run-time contexts. So, the former `Syntax.equal_binding`
function is now `syntax_meta.equal_binding`, and it is provided by
`rhombus/meta` instead of `rhombus`. (A functoon like
`free-identifier=?` is used in two different ways in Racket: by a
macro to decide how to expand, or by a tool that processes expanded
code to inspect bindings. The latter doesn't really apply in Rhombus,
so we focus on the former and put it in `rhombus/meta`.) The
`Syntax.meta_value` function is renamed `syntax_meta.value`. To keep
all the compile-time syntax operations together, `Syntax.error` is now
`syntax_meta.error`.

To use a function like `syntax_meta.equal_binding` or
`syntax_meta.value` for a space other than the expression space, some
compile-time value needs to refer to the space. So, there are now
things like `expr_meta.space`, `bind_meta.space`, and the `reflection`
meta-clause for the `space` form.

Finally, `Term`, `Group`, `Name`, and `IdentifierName` are now all
annotations as well as syntax classes. For example,
`Syntax.meta_value` expects a value satisfying `Name` as its first
argument.