add `with` for functional update of objects
Example:

 class Posn(x, y)
 Posn(1, 2) with (x = 10)
 // prints `Posn(10, 2)`

The general syntax is `obj_expr with (field_id = expr, ...)`, where
each `field_id` is a public field of the object that is represented in
the constructor. A new object is generated by calling a
"reconstructor" for the class that the object instantiates; by
default, the reconstructor calls the class's constructor.

In the case that static information is available for the target
object, the update can be more efficient. For a final class, an update
expands to a direct call to the constructor. For a non-final class, an
update is the same as a dispatch to a method that may have optional
arguments with defaults drawn from the target object's fields.

If no static information is available, then a dynamic lookup is used.
In this case, the fields must be comparsed against the class's field
list at run time, and an `apply` is eventully used to call the
reconstrtcutor. No keywords are involved, however, since all fields
are supplied to a reconstructor by position.

A `class` declaration can specialize update with a `reconstructor`
clause, simialr to writing a method.