add `where` for binding in the process of matching
Like `when`, the `where` operator creates a side condition on a
binding pattern match. Unlike `when`, `where` binds names that are
visible in the same scope as the names bound by the left-hand pattern.
A match successed only if both the main pattern and its `where`
bindings all match.

Besides the addition of `where`, add support for splicing of
multi-group syntax objects in place of a whole-group context, and
clean up some list functions.