static `for`: check for sequences (possibly specialized)
A static `for` now checks that the right-hand side of an `each` is
known to be a sequence. If the right-hand side simply has `Sequence`
as its static information, that doesn't do anything, but the
expectation is that a more specific annotation will be available to
enable specialization of the iteration.