fill out context parameters and add syntax parameters
Instead of a wrapping form like Racket's `syntax-parameterize` or
`splicing-syntax-parametetrize`, a syntax parameter here is updated
with `syntax_parameter.relet`, and it applies to all subsequent form
in the same definition context. This difference is in keeping with the
general avoidance of partial expansion and rewriting partially
expanded terms in Rhombus. Instead, syntax parameters are implemented
with the same expansion driver that handles `let`, although that
requires more cooperation with forms like `class` that implement their
own definition contexts.