move definition and declaration bindings to their own spaces
Formerly, definitions via `defn.macro` and declarations via
`decl.macro` used the same space as expression bindings. That's
convenient in some ways, but it doesn't allow binding an identifier as
say, both a definition and expression form, in the same way that `fun`
was primitively defined. Shifting definitions and declarations to
their own spaces brings the convention in line with other kinds of
bidnings.

This split requires a little more work to get documentation and
typesetting right in Scribble, but we simplify the common case by
adding a fallback to check the declaration and definition space by
default for label bindings.